requirements.txt
loguru==0.7.3
google-genai>=1.21.1
